Extract wait strategy tests to own files
